requirements.txt
django==4.0.5
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangp-cors-headers==3.13.0